+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [Unreleased]
+
+### Changed
+- Refactor InitialSetup pane to use compose instead of set_data + mount
+- Refactor Db4E to use compose instead of set_data + mount
+
+---
+
 ## [0.23.1] - 2025-07-15
 
 ### Added